Update index.html
The contents below the profile picture, pete's background and contact section needs to be aligned towards the center.
Navigation links are not working; they should take you to the correct section of the webpage when clicked.
The whole 'contact me' word here - Please feel free to take a look at my website and feel free to contact me. should be a link, when clicked should take to the contact section of the webpage. Please correct these errors.